refactor: Create more shared classes for form elements
This commit includes the creation of shared (or just nested selectors) of the follwing:
- <h2> as 'main h2' for major sections
- <h3> as 'main h3' for subsections
- <h4> as 'main h4' for parts within subsections
The commit also comments so I don't forget.

The use of descendants is better, imo, in this case because I've already mapped out which headings to use where, and it keeps the HTML squeaky clean.
Note: upload mockup images to README, or provide Figma link