wip: start implementing dynamic add/remove dropdowns for Disciplines/Backgrounds/Merits/Flaws
This commit starts the process of modifying the HTML structure (adding ids) and adding logic to the app.js.

Right now, it's a work in progress. Adding dropdowns work, and they are being populated. However, it is appearing *below* the add-btn, when it should be appearing *above* it.
The new dropdowns also do not have the placeholder text. Additionally, adding a new dropdown will reset selected options in previous dropdowns. No, no.
Merit dropdowns also aren't working.