fix: fix Merit dropdowns not removing, and new dropdowns appearing below the add button
This commit changes the `id="merit-flaw-wrapper"` to `class="merit-flaw-wrapper"`. It was supposed to be a class, NOT id, OOPS!
For the layout issue, a line in the `addButton` eventListener inside `initializeDynamicRows` is changed from `rowContainer.appendChild(newRow);` to `rowContainer.insertBefore(newRow, addButton.parentElement);`.

Fixing the placeholder issue will be next!