fix: Prevent dropdown reset on new row creation
The previous implementation caused existing dropdowns (specifically Disciplines and Backgrounds) to have their selected values reset/cleared whenever a new dropdown/row was added.

This was because there was an argument mismatch in the generic callback, which caused the populate function to re-render all dropdowns instead of just the new one. THe callback is now more explicity, making sure only the newly created element is affected.