fix: Add space for `.btn-minus` to stop the Jiggle
The layout no longer moves or Jiggles when the Merit/Flaw dropdown 'remove button' appears when the screen size is `min-width:640px` or `sm`.

`.btn-minus` was modified so that space is preserved for it, and so the UI doesn't need to compensate for its sudden appearance. Additionally, only the opacity and the scale is animated, not the actual slide. So it just grows and shrinks instead of sliding out. Aw.
The new dropdowns are also centered with `ml-10`.