fix: Fill-in 1 starting dot for Virtues
The `initializeSimpleDotLogic` function was missing the initial setup call  and the necessary logic for the starting and permanent Virtue dots.

So added initial dot setup where a loop goes through every `.dot-group` in the section and explicitly sets the `filled` class based on the `baseDotsPerItem` value.

Also added a new check inside the `updateDotsInGroup` helper function. Should act like the Attributes section now.