style: revamp freebie mode button
This commit modifies the freebie mode button so that it's more in-line with the new plan for freebie mode.

There's now two freebie points buttons. One is the button to toggle/enter the freebie points mode and the second is actually a div that displays the available freepoints as well as a button to reset all freebie point distribution.

The second button is hidden be default and would only appear once the user has entered freebie mode. At that point, the first button will then be hidden.