fix: make freebie dot handler more explicit
This commit adds `data-section-id` tags to all sections (excluding Merits/Flaws) to make it more explicit for the freebie dot handling functions to target.

The `initializeFreebieListeners` function now doesn't care what element it's looking for.  `dot.closest('[data-section-id]')` just needs to find the closest dot to the element that has the `data-section-id`.

No more issues with the html structure. hopefully.